requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
Pillow==11.2.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-django==4.11.1
